+++ src/components/Footer.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Footer.scss';
+
+const Footer = () => {
+  return (
+    <footer>
+      <p>Copyright&copy; 2024 the13thgeek.<br />All rights reserved.</p>
+    </footer>
+  )
+}
+
+export default Footer